+++ src/lib/index.es2015.ts
@@ -1,3 +1,5 @@
+// This file assumes that the environment supports ES2015 (ES6)
+
 export {Type, SerializableType, VersionedType} from "./interfaces";
 
 export {ArrayType} from "./types/array";
@@ -7,10 +9,12 @@ export {DateType} from "./types/date";
 export {DocumentType} from "./types/document";
 export {Float64Type} from "./types/float64";
 export {Int32Type} from "./types/int32";
+export {LiteralType} from "./types/literal";
 export {NullType} from "./types/null";
 export {SimpleEnumType} from "./types/simple-enum";
 export {TypedUnionType} from "./types/typed-union";
 export {Ucs2StringType} from "./types/ucs2-string";
+export {UnionType} from "./types/union";
 
 import * as array from "./types/array";
 import * as boolean from "./types/boolean";
@@ -19,10 +23,12 @@ import * as date  from "./types/date";
 import * as document  from "./types/document";
 import * as float64  from "./types/float64";
 import * as int32  from "./types/int32";
+import * as literal  from "./types/literal";
 import * as _null  from "./types/null";
 import * as simpleEnum  from "./types/simple-enum";
 import * as typedUnion  from "./types/typed-union";
 import * as ucs2String  from "./types/ucs2-string";
+import * as union  from "./types/union";
 export {
   array,
   boolean,
@@ -31,10 +37,12 @@ export {
   document,
   float64,
   int32,
+  literal,
   _null as null,
   simpleEnum,
   typedUnion,
-  ucs2String
+  ucs2String,
+  union
 };
 
 import * as errors from "./errors/index";